Implement CRUD for extra compensatory leaves
There is a problem of communication between FosRest and EmberData.
FosRest does not allows automatically set the api paths like
/api/compensatoryLeaves (that Ember requires)
but only
/api/compensatoryleaves (lowercased)

To solve the problem I use a redirect rule in web.config. But it doesn't work
for POST request. Because of it I add for only for POST method of
API-controller manual a path's correction.

I'm not happy with the solution :/